ll option added

+++ removeTaxa.py
@@ -10,13 +10,13 @@ if len(sys.argv) == 4:
 	files = glob.glob(inputfolder+"/*.fas")
 	if sys.argv[2] == "-a" or sys.argv[2] == "-ar" or sys.argv[2] == "-e":
 		exclusion_file = sys.argv[3]
-	elif sys.argv[2] == "-l":
+	elif sys.argv[2] == "-l" or sys.argv[2] == "-ll":
 		threshold = float(sys.argv[3])
 	else:
 		print "incorrect command line parameters"
 		sys.exit()
 else:
-	print "FORMAT: python removeTaxa.py [folder with fasta] [option: -a (leave specified taxa), -ar (leave and rename specified taxa), -e (exclude specified taxa), -l (exclude short seq taxa)] [taxalist (or csv) or lenght percent threshold]"
+	print "FORMAT: python removeTaxa.py [folder with fasta] [option: -a (leave specified taxa), -ar (leave and rename specified taxa), -e (exclude specified taxa), -l (exclude short seq taxa), -ll (exclude loci with few taxa)] [taxalist (or csv) or lenght percent threshold]"
 	print "EXAMPLE: python removeTaxa.py ./fasta -l 0.75"
 	print "EXAMPLE: python removeTaxa.py ./fasta -a list.lst"
 	sys.exit()
@@ -55,23 +55,28 @@ for f in files:
 	prog = "working on file "+fn
  	sys.stdout.write(prog+"\r")
  	sys.stdout.flush()
-	outputfile=open("./reduced/"+fn, "w")
-	count = 0
-	for seq in SeqIO.parse(f, "fasta"):
-		#print seq.id, len(str(seq.seq).replace("-", "").replace("N", "")), len(seq.seq), float(len(str(seq.seq).replace("-", "").replace("N", "")))/len(seq.seq)
-		if sys.argv[2] == "-e" and seq.id not in exclusion_list:
-			print >> outputfile, ">"+seq.id, "\n", seq.seq
-			count += 1
-		elif sys.argv[2] == "-a" and seq.id in exclusion_list:
-			print >> outputfile, ">"+seq.id, "\n", seq.seq
-			count += 1
-		elif sys.argv[2] == "-ar" and seq.id in exclusion_list:
-			print >> outputfile, ">"+exclusion_list[seq.id], "\n", seq.seq
-			count += 1
-		elif sys.argv[2] == "-l" and float(len(str(seq.seq).replace("-", "").upper().replace("N", "")))/len(seq.seq)>threshold:
-			print >> outputfile, ">"+seq.id, "\n", seq.seq
-			count += 1
-	outputfile.close()
-	if count == 0:
-		os.remove("./reduced/"+fn)
+ 	if sys.argv[2] == "-ll":
+ 		#print len([SeqIO.parse(f, "fasta")])
+ 		if len(list(SeqIO.parse(f, "fasta"))) >= threshold:
+ 			shutil.copy2(inputfolder+fn, "./reduced")
+ 	else:
+		outputfile=open("./reduced/"+fn, "w")
+		count = 0
+		for seq in SeqIO.parse(f, "fasta"):
+			#print seq.id, len(str(seq.seq).replace("-", "").replace("N", "")), len(seq.seq), float(len(str(seq.seq).replace("-", "").replace("N", "")))/len(seq.seq)
+			if sys.argv[2] == "-e" and seq.id not in exclusion_list:
+				print >> outputfile, ">"+seq.id, "\n", seq.seq
+				count += 1
+			elif sys.argv[2] == "-a" and seq.id in exclusion_list:
+				print >> outputfile, ">"+seq.id, "\n", seq.seq
+				count += 1
+			elif sys.argv[2] == "-ar" and seq.id in exclusion_list:
+				print >> outputfile, ">"+exclusion_list[seq.id], "\n", seq.seq
+				count += 1
+			elif sys.argv[2] == "-l" and float(len(str(seq.seq).replace("-", "").upper().replace("N", "")))/len(seq.seq)>threshold:
+				print >> outputfile, ">"+seq.id, "\n", seq.seq
+				count += 1
+		outputfile.close()
+		if count == 0:
+			os.remove("./reduced/"+fn)
 print "\ndone"
